projects/voice_control/requirements.txt
SpeechRecognition==3.10.0
paramiko==3.4.0
pyttsx3==2.90
pywhatkit==5.4
streamlit==1.31.1
gradio==5.34.1
numpy==1.26.4
python-dotenv==1.0.1
pyaudio==0.2.14
pyautogui==0.9.54
pywin32==306
colorama==0.4.6 